pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "blendpy"
version = "25.11.0"
authors = [
  { name="Leandro Seixas", email="leandro.seixas@proton.me" },
]
description = "Blendpy is a computational toolkit for investigating thermodynamic models of alloys using first-principles calculations."
readme = "README.md"
license = {file = "LICENSE"} 
requires-python = ">=3.8"
dependencies = [
    'numpy',
    'pandas',
    'ase',
    'pytest',
    'pyyaml',
    'tqdm',
    'mpltern'
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Physics"
]

[project.urls]
"Homepage" = "https://github.com/leseixas/blendpy"
"Repository"= "https://github.com/leseixas/blendpy"

